move MMKVZustandStorage to App
