+++ src/frontend/App.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import {getLocales} from 'expo-localization';
 import * as SecureStore from 'expo-secure-store';
 import {createMapeoClient} from '@comapeo/ipc';
+import {type StateStorage} from 'zustand/middleware';
 import {AppNavigator} from './AppNavigator';
 import {MessagePortLike} from './lib/MessagePortLike';
 import {initializeNodejs} from './initializeNodejs';
@@ -14,10 +15,7 @@ import * as Sentry from '@sentry/react-native';
 import * as TaskManager from 'expo-task-manager';
 import {applicationId} from 'expo-application';
 import {LOCATION_TASK_NAME, LocationCallbackInfo} from './sharedTypes/location';
-import {
-  MMKVZustandStorage,
-  storage,
-} from './hooks/persistedState/createPersistedState';
+import {storage} from './hooks/persistedState/createPersistedState';
 import {useOnBackgroundedAndForegrounded} from './hooks/useOnBackgroundedAndForegrounded';
 import {getSentryUserId} from './metrics/getSentryUserId';
 import {AppDiagnosticMetrics} from './metrics/AppDiagnosticMetrics';
@@ -52,6 +50,21 @@ Sentry.init({
 });
 
 Mapbox.setTelemetryEnabled(false);
+
+// TODO: No longer need to export this when the updated implementation of the draft observation hook is used
+// (`hooks/persistedState/createPersistedState.ts` will be removed)
+export const MMKVZustandStorage: StateStorage = {
+  setItem: (name, value) => {
+    return storage.set(name, value);
+  },
+  getItem: name => {
+    const value = storage.getString(name);
+    return value ?? null;
+  },
+  removeItem: name => {
+    return storage.delete(name);
+  },
+};
 
 const persistedLocaleStore = createLocaleStore({
   persist: true,
